% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{bestmatch}
\alias{bestmatch}
\title{Best string match using string distance}
\usage{
bestmatch(string, stringVector)
}
\arguments{
\item{string}{elements to be approximately matched: will be coerced to
\code{character} unless it is a list consisting of \code{integer} vectors.
Identical to the \code{x} parameter in \code{stringdist::amatch()}}

\item{stringVector}{a vector of strings to be used as lookup table for
matching. Will be coerced to \code{character} unless it is a list consisting
 of \code{integer} vectors. Identical to the \code{table} parameter in
 \code{stringdist::amatch()}}
}
\value{
a vector of the same length as \code{string} containing the position
of the closest match of each element of \code{string} in \code{stringVector}.
Identical to the result returned by \code{stringdist::amatch()}.
}
\description{
Find the best match for one character string in a vector of character strings
using Levenshtein distance.
}
\details{
The function \code{bestmatch} is a quick wrapper around the \code{stringdist::amatch()}
function using \code{method = "lv"} and \code{maxDist = 1}
}
