% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{fill.blanks}
\alias{fill.blanks}
\title{"Fill down the blank"}
\usage{
fill.blanks(x)
}
\arguments{
\item{x}{A pre-sorted vector that may contain \code{NA} values or empty strings.}
}
\value{
A vector of the same length and type as \code{x}, with all \code{NA}
values or empty strings replaced by the value of the preceding element in the
vector.
}
\description{
For a pre-sorted vector, fill each missing value with that of the preceding
element.
}
\details{
\code{fill.blanks()} is a wrapper around the \code{zoo::na.locf()} function
in the package \code{zoo}. It can handle both missing values from all data
types including character vectors where missing values are often encoded as
empty strings.
}
