% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{find.numeric}
\alias{find.numeric}
\title{Identify numeric values from a character vector}
\usage{
find.numeric(x, response = c("index", "value", "boolean"))
}
\arguments{
\item{x}{A character vector containing possible numeric values that need
identified.}

\item{response}{The type of values to be returned. One of "index" (default),
"value", or "boolean".}
}
\value{
If \code{response == "index"}, the indices of all numeric values. If
  \code{response == "value"}, the values of these numeric values, with
  non-numeric values replaced by \code{NA}. If \code{response == "boolean"},
  boolean indicators for whether each value in \code{x} is numeric or string.
}
\description{
Identify all numbers from a character vector (e.g. ID variables). Can return
indices of these numbers, their values, or boolean indicators for whether
each value is numeric.
}
