% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{genperms}
\alias{genperms}
\title{Generate flexible permutation matrix for blocked, clustered or simpler
designs}
\usage{
genperms(x, block = NULL, clus = NULL, maxiter = 10000)
}
\arguments{
\item{x}{a vector to be permuted. Can be continuous}

\item{block}{a vector of equal length as \code{x}, with unique values
indicating different blocks}

\item{clus}{a vector of equal length as \code{x}, with unique values
indicating different blocks}

\item{maxiter}{maximum number of permutations to be included in the
permutation matrix. If it is possible to perform exact permutation with
fewer permutations, then the exact permutation matrix is produced.}
}
\value{
A permutation matrix where each row is a permutation of the input
  vector \code{x}
}
\description{
An alternative version of \code{ri::genperms}. Given user-input vectors of
clusters or blocks, generate an exact permutation matrix, or a randomly
sampled permutation matrix if the number of actual permutations is too high.
Improves upon \code{ri::genperms} by allowing permutation of non-binary
vectors.
}
\details{
Unlike its counterpart in the \code{ri} package, this function can perform
permutation of input vectors that are not binary. It also accepts as
arguments for \code{blockvar} and \code{clustvar} other data types than
\code{integer}. \code{genperms} is primarily based on the \code{permute}
package.
}
\references{
Gerber, Alan S. and Donald P. Green. 2012. Field Experiments:
  Design, Analysis, and Interpretation. New York: W.W. Norton.
}
