% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{neglog}
\alias{neglog}
\title{Negative-log transformation}
\usage{
neglog(x, base = exp(1), offset = 0.5)
}
\arguments{
\item{x}{a numeric or complex vector}

\item{base}{a positive or complex number: the base with respect
to which logarithms are computed. Defaults to \eqn{e=}\code{exp(1)}}

\item{offset}{a number specifying an offset to avoid \code{neglog(x)}
returning \code{-Inf}. Defaults to 0.5}
}
\value{
A vector of the same length as \code{x} containing the transformed
values.
}
\description{
An analogue to the log transformation for negative values.
}
\details{
\code{neglog(x)} is calculated such that
\eqn{neglog(x) = sign(x) \times \log |x + offset|}{neglog(x) = sign(x) * \log(|x + offset|)}.
The offset ensures that no \code{-Inf} value is returned.
The default offset is 0.5.
}
