% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{papi}
\alias{papi}
\title{etnam PAPI Data}
\format{An object of class \code{data.frame} with 378 rows and 32 columns.}
\usage{
papi
}
\description{
A data frame containing results from the Public Administration Performance
  Index (PAPI) survey, aggregated at the province level for all Vietnamese
  provinces from 2005 to 2016. It contains 15 variables, which include
  province and year IDs, weighted and unweighted aggregate PAPI scores, and
  component PAPI scores.
}
\details{
Raw data for \code{pci} is scraped from the website of the Vietnam
  Provincial Governance and Public Administrataion Performance Index Project
  at \url{papi.org.vn}
}
\keyword{datasets}
