% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pci}
\alias{pci}
\title{Vietnam PCI Data}
\format{An object of class \code{data.frame} with 738 rows and 15 columns.}
\usage{
pci
}
\description{
A data frame containing results from the Provincial Competitiveness Index
  (PCI) survey, aggregated at the province level for all Vietnamese provinces
  from 2005 to 2016. It contains 15 variables, which include province and
  year IDs, weighted and unweighted aggregate PCI scores, and component PCI
  scores.
}
\details{
Raw data for \code{pci} is scraped from the website of the Provincial
  Competitiveness Index Project at \url{www.pcivietnam.org}
}
\keyword{datasets}
