% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{plot.riFit}
\alias{plot.riFit}
\title{Plot Point Estimate and Randomization Distribution for Randomization
Inference Result}
\usage{
\method{plot}{riFit}(x, title = NULL, xlab = NULL, ylab = NULL,
  scale = F, xmin, xmax, axe.y = F, ...)
}
\arguments{
\item{x}{An object of the class \code{riFit} generated by
\code{rireg} or \code{riwfe}.}

\item{title}{The main title (on top)}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{scale}{logical; if TRUE the point estimates and the randomization
distribution is demeaned and normalized before plotting (default is FALSE)}

\item{xmin}{a numeric indicating the minimum value of the plot's main axix}

\item{xmax}{a numeric indicating the maximum value of the plot's main axix}

\item{axe.y}{logical; if TRUE the y-axis is shown (default is FALSE)}

\item{...}{Other arguments passed on to \code{ggplot2::ggplot}. Currently unused}
}
\description{
A plot showing the point estimate for the treatment effect as a red line, and
the randomization distribution of the treatment effect as gray bars.
}
