% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{plot.riSynth}
\alias{plot.riSynth}
\title{Plot Point Estimate and Randomization Distribution for Randomization
Inference Result from riSynth}
\usage{
\method{plot}{riSynth}(x, title = NULL, xlab = NULL, ylab = NULL,
  scale = F, xmin, xmax, att = T, axe.y = F)
}
\arguments{
\item{x}{an \code{riSynth} object}

\item{title}{The main title (on top)}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{scale}{logical; if TRUE the point estimates and the randomization
distribution is demeaned and normalized before plotting (default is FALSE)}

\item{xmin}{a numeric indicating the minimum value of the plot's main axix}

\item{xmax}{a numeric indicating the maximum value of the plot's main axix}

\item{att}{logical; currently unused}

\item{axe.y}{logical; if TRUE the y-axis is shown (default is FALSE)}
}
\description{
A plot showing the point estimate for the treatment effect as a red line, and
the randomization distribution of the treatment effect as gray bars.
}
\details{
\code{plot.riSynth} converts a \code{riSynth} object into a \code{riFit} object
and then make a call to \code{plot.riFit}.
}
