% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{rireg}
\alias{rireg}
\title{Randomization Inference on Treatment Effect using Linear Regression}
\usage{
rireg(data, outcome, treatment, covs, perm = NULL, blockvar = NULL,
  clustvar = NULL, maxiter = 10000, covs_control = list(method =
  "Euclidean", tol_quantile = 0.05, tol_function = function(x) 0.25 * sd(x),
  tol_value = NULL))
}
\arguments{
\item{data}{a data frame containing the variables in the model}

\item{outcome}{a character. Name of the outcome variable.}

\item{treatment}{a character. Name of the treatment variable.}

\item{covs}{a character vector. Names of the covariates to be used in the
model.}

\item{perm}{a matrix containing permutations of the treatment variable (for
\code{rireg} or \code{riSynth}) or the outcome variable (for \code{riwfe}).
When \code{perm} is supplied the call to \code{genperms} will not be made
and the arguments \code{blockvar}, \code{clustvar} and \code{maxiter} will
be ignored}

\item{blockvar}{an optional character vector. Name of the block variable if
the randomization inference procedure requires block randomization. The
variable named by \code{blockvar} will be used as input for the
\code{genperms} function.}

\item{clustvar}{an optional character vector. Name of the cluster variable if
the randomization inference procedure requires clustered randomization. The
variable named by \code{clustvar} will be used as input for the
\code{genperms} function.}

\item{maxiter}{a positive integer. The maximum number of permutations to be
included in the permutation matrix for the randomization distribution. Used
as input for the \code{genperms} function.}
}
\value{
An object of class \code{riFit}
}
\description{
Estimates of the treatment effect using linear models through \code{lm}, then
conducts inference using randomization inference by permuting the treatment
vector to obtain the sharp null distribution
}
\details{
Estimates of the treatment effects are obtained by OLS regression. When
covariates are included, the randomization distribution is obtained by
permuting the "partialled-out" treatment vector i.e. the vector of residuals
from a regression of treatment on covariates. Internally \code{rireg} makes
call to \code{genperms}. The variable whose names are given by
\code{blockvar} and \code{clustvar} will be coerced into input vectors for
the \code{block} and \code{clus} arguments of the \code{genperms} function.
}
