% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{riSynthToriFit}
\alias{riSynthToriFit}
\title{Convert \code{riSynth} object to \code{riFit}}
\usage{
riSynthToriFit(riSynth.obj)
}
\arguments{
\item{riSynth.obj}{An object of the class \code{riSynth} generated by
\code{riSynth}.}
}
\value{
a \code{riFit} object
}
\description{
Convert a \code{riSynth} to a \code{riFit} object
}
\details{
Useful for plotting. Converts all \code{tau.att} elements in the \code{riSynth}
object to corresponding \code{beta} elements in \code{riFit} and drops all
\code{tau.i} elements.
}
