% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{riwfe}
\alias{riwfe}
\title{Randomization Inference on Treatment Effect using Weighted Fixed Effects
Regression}
\usage{
riwfe(data, outcome, treatment, covs, perm = NULL, blockvar = NULL,
  clustvar = NULL, maxiter = 1000, unit.index, time.index, method,
  qoi = "ate", estimator = NULL, unbiased.se = TRUE,
  covs_control = list(method = "Euclidean", tol_quantile = 0.05, tol_function
  = function(x) 0.25 * sd(x), tol_value = NULL))
}
\arguments{
\item{data}{a data frame containing the variables in the model}

\item{outcome}{a character. Name of the outcome variable.}

\item{treatment}{a character. Name of the treatment variable.}

\item{covs}{a character vector. Names of the covariates to be used in the
model.}

\item{perm}{a matrix containing permutations of the treatment variable (for
\code{rireg} or \code{riSynth}) or the outcome variable (for \code{riwfe}).
When \code{perm} is supplied the call to \code{genperms} will not be made
and the arguments \code{blockvar}, \code{clustvar} and \code{maxiter} will
be ignored}

\item{blockvar}{an optional character vector. Name of the block variable if
the randomization inference procedure requires block randomization. The
variable named by \code{blockvar} will be used as input for the
\code{genperms} function.}

\item{clustvar}{an optional character vector. Name of the cluster variable if
the randomization inference procedure requires clustered randomization. The
variable named by \code{clustvar} will be used as input for the
\code{genperms} function.}

\item{maxiter}{a positive integer. The maximum number of permutations to be
included in the permutation matrix for the randomization distribution. Used
as input for the \code{genperms} function.}

\item{unit.index}{a character string indicating the name of unit
    variable used in the models. The index of unit should be factor.
  }

\item{time.index}{a character string indicating the name of time
    variable used in the models. The index of time should be factor.
  }

\item{method}{method for weighted fixed effects regression, either
    \code{unit} for unit fixed effects; \code{time} for time fixed
    effects. The default is \code{unit}. For two-way weighted fixed
    effects regression models, set method to the default value
    \code{unit}.}

\item{qoi}{one of \code{"ate"} or \code{"att"}. The default is
    \code{"ate"}. If set to \code{"att"} in implementing \code{"fd"}
    and \code{"did"} estimators, the comparison of the treated
    observation is restricted to the control observation from the
    previous time period but not with the control observation from the
    next time period.}

\item{estimator}{an optional character string indicating the
  estimating method. One of \code{"fd"}, \code{"did"}, or
  \code{"Mdid"}. \code{"fd"} is for First-Difference
  Design. \code{"did"} is for multi-period Difference-in-Differences
  design. The default is \code{NULL}. Setting estimator to be
  \code{"Mdid"} implements the Difference-in-Differences design with
  Matching on the pretreatment outcome variables.}

\item{unbiased.se}{logical. If \code{TRUE}, bias-asjusted
    heteroskedasticity-robust standard errors are used. See Stock and
    Watson (2008). Should be used only for balanced panel. The default
    is \code{FALSE}.
  }
}
\value{
An object of class \code{riFit}
}
\description{
Estimates of the treatment effect using Weighted Fixed Effects Regression
through the \code{wfe} function in the \code{wfe} package, then conducts
inference using randomization inference by permuting the treatment vector to
obtain the sharp null distribution
}
\details{
Estimates of the treatment effects are obtained by OLS regression. When
covariates are included, the randomization distribution is obtained by
permuting the outcome vector. This is equivalent to permuting the treatment
vector and all associated covariates. Unlike \code{rireg}, the \code{riwfe}
function does not make use of the partialling-out method because the function
\code{wfe} from the \code{wfe} package that \code{riwfe} calls does not allow
the treatment variable to be omitted. Internally, \code{riwfe} makes call to
\code{genperms}. The variable whose names are given by \code{blockvar} and
\code{clustvar} will be coerced into input vectors for the \code{block} and
\code{clus} arguments of the \code{genperms} function. The arguments
\code{unit.index}, \code{time.index}, \code{method}, \code{qoi},
\code{estimator}, \code{unbiased.se} are input directly into the call for
\code{wfe}.
}
