% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{SynthATT}
\alias{SynthATT}
\title{Synthetic Control ATT}
\usage{
SynthATT(data, outcome, treatment, covs, treatment.year, pretreatment.year,
  posttreatment.year = NULL, unit.variable, unit.names.variable,
  time.variable, include.past.Y = TRUE, snowfall = FALSE)
}
\arguments{
\item{data}{a data frame containing the variables in the model}

\item{outcome}{a character. Name of the outcome variable.}

\item{treatment}{a character. Name of the treatment variable.}

\item{covs}{a character vector. Names of the covariates to be used in the
model.}

\item{treatment.year}{the year or time period at which treatment occurs}

\item{pretreatment.year}{the years or time periods before treatment occurs. Observations from these
years or time periods will be used to create the synthetic contorl}

\item{posttreatment.year}{the years or time periods after treatment occurs. Observations form these
years or time periods will be used to calculate treatment effects.}

\item{unit.variable}{A scalar identifying the column number or column-name character string associated unit numbers. The unit.varibale has to be numeric.}

\item{unit.names.variable}{A scalar or column-name character string identifying the column with the names of the units. This variable has to be of mode character. 
  }

\item{time.variable}{A scalar identifying column number or column-name character string associated with period (time) data. The time variable has to be numeric.}

\item{include.past.Y}{a logical; if TRUE past values of the outcomes are included as a covariate
to produce the synthetic control}

\item{snowfall}{a logical; if TRUE the function performs parallel computing
using the snowfall package}
}
\value{
An list with two objects: \code{tau.i}, a vector of unit-specific treatment effects, and
\code{tau.att} the Average Treatment Effect on the Treated
}
\description{
Estimates the average treatment effect on the treated using the Synthetic
Control Method, then conducts
inference using randomization inference by permuting the treatment vector to
obtain the sharp null distribution
}
\details{
For each unit in the treatment group, \code{Synthatt} creates a synthetic
control using weighted averages of the units in the control group, then
estimates an unit-specific treatment effect. The Average Treatment Effect on
the Treated is calculated by taking the average of all unit-specific
treatment effects.

Internally, \code{Synthatt} makes use of the \code{synth} and \code{dataprep}
functions in the \code{Synth} package. For computers with multiple cores,
\code{Synthatt} can perform parallel computation to improve computation
speed.
}
