% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VNbudget_final}
\alias{VNbudget_final}
\title{Vietnam Provincial Budget Finals}
\format{An object of class \code{data.frame} with 569 rows and 37 columns.}
\usage{
VNbudget_final
}
\description{
A data frame containing realized national budget broken down by provinces
  for all Vietnamese provinces from 2006 to 2014. Realized budgets are
  calculated at the end of every year. It contains 38 variables, which
  include province and year IDs, total revenues, total expenditures, central
  transfers broken down by categories, as well as log transformations, lags
  and change values for the latter.
}
\details{
Raw data for \code{VNbudget_final} is scraped from the website of the
  Vietnamese Ministry of Finance at \url{www.mof.gov.vn}
}
\keyword{datasets}
