% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VNbudget_plan}
\alias{VNbudget_plan}
\title{Vietnam Provincial Budget Plans}
\format{An object of class \code{data.frame} with 759 rows and 46 columns.}
\usage{
VNbudget_plan
}
\description{
A data frame containing planned national budget broken down by provinces
  for all Vietnamese provinces from 2006 to 2016. Planned budgets are issued
  at the beginning of every year. It contains 46 variables, which include
  province and year IDs, total revenues, total expenditures, central
  transfers broken down by categories, as well as log transformations, lags
  and change values for the latter.
}
\details{
Raw data for \code{VNbudget_plan} is scraped from the website of the
  Vietnamese Ministry of Finance at \url{www.mof.gov.vn}
}
\keyword{datasets}
