% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{year}
\alias{year}
\title{Extract year from date}
\usage{
year(date)
}
\arguments{
\item{date}{a \code{character} object or vector indicating dates in
one of three formats: exact date in d/m/Y, exact year without date,
and date as number of days since 01/01/1900.}
}
\value{
The year of the date in \code{date}
}
\description{
Extract the year from a date variable in multiple formats
}
\details{
\code{year} can handle dates in the following formats: date in
d/m/Y, year without date, and date as number of days since 01/01/1900.
}
